requirements.txt
Django>=4.2,<5.0
python-dotenv
Pillow

# IA (serão usados mais adiante)
numpy
torch
tensorflow

# Testing
pytest
pytest-django
pytest-playwright